Comparison of job position vs posting key words
A duplication of parsing a job position for key words (list tuple) and a job posting for similarities. Not an elegant solution but handy to continue building awareness in a general audience of relevance of Data Science/Python in ordinary tasks. Key words as a building block of the strategy to understand the digital shift in organizations and those aspiring to join them or running them ,i.e. Executive Directors. Closing the digital gap one python code line at a time.